tests/wpt/web-platform-tests/content-security-policy/navigate-to/meta-refresh-blocked.html

<!DOCTYPE html>

<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>

<body>
<script>
  var t = async_test("Test that the child iframe navigation is blocked");
  window.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data.result, 'fail');
    assert_equals(e.data.violatedDirective, 'navigate-to');
  });
</script>
<iframe src="support/meta_refresh_navigation.sub.html?csp=navigate-to%20%27none%27&target=post_message_to_frame_owner.html">
</body>
